Fix issue with eval on JRuby.
In the specs, when running for the first time, all values like StringVal,
IntVal, etc. are defined properly. But once you unset the `Mod` module
from Object, and try to re-define StringVal, it would not find Mod::StringVal
anymore. Using `module_eval` instead of `eval` to keep the context of the
module fixes it.