Modify RoleValues to have the same mechanism as Instance Index.
RoleValues was basically a wrapper for an array, and is now similar to an ordered
hash. The basic functionnality of RoleValues have not been modified, and it should still
be considered as an array on multiple points, but this array will be ordered by its values
automatically.

Refactoring to InstanceIndexKey has been done also to be used by RoleValues. Therefore, its
name has been modified to be ComparableHashKey instead.