Fix tests for multipart_identification_spec.
Adds information into related_entities. It should contain a RoleValue
object when the role is pointing to many entities (see unique attribute
for roles).

It also adds a way to query entities in a two-way fashion. Which means
you can now search for entities which references the current one, but
also entities which are referenced by the current one.